Error handling for combined multithreading
The tile parsing errors are already handled. The only error that
has to be handled is the errors in the tile decoding. This CL
handles that.

The overall frame parallel error path remains the same (using
the SignalFailure() function). This CL makes sure that the
DecodeTilesThreadedFrameParallel() always returns the right error
code on failure.

PiperOrigin-RevId: 303856964
Change-Id: I62faa43bf09b10ed65c45a761b93734a4159bc7f