update RefCountedBuffer::SetFrameDimensions()
Reset motion_field_reference_frame_ and motion_field_mv_ only when necessary.
They are set in Tile::StoreMotionFieldMvsIntoCurrentFrame() and read in
MotionFieldProjection(). Both functions are not called (or early terminated)
for intra frames or when refresh_frame_flags is 0.

Set zero_initialize parameter of motion_field_reference_frame_.Reset() to true
to do the initialization.

PiperOrigin-RevId: 304302322
Change-Id: I779a9585680e858c1361ab6943c8e22ebcef4087