Make kMaxThreads configurable at compile time
Since the introduction of frame parallel mode, applications may
choose to compile libgav1 with a larger thread count limit. Allow
that as an option.

Also increase the default value from 32 to 128.

The memory consumption doesn't change drastically with the max
thread limit since we only use it to create the stack for
TileScratchBufferPool and FrameScratchBufferPool which are just
arrays of std::unique_pointers.

PiperOrigin-RevId: 304718347
Change-Id: I7b604a58cd4dc284339b89f6e788ca14c214baa5