Do not hold lock when notifying condition variables
Notifying the condition variables does not have to be synchronized. So
it can be outside the critical section where ever possible.

Also remove one of the decoded_condvar_.notify() calls from
ParseAndSchedule() since there can be no waiters in DequeueFrame() when
ParseAndSchedule() is running (the API contract mandates that the
application cannot call EnqueueFrame and DequeueFrame at the same time).

PiperOrigin-RevId: 305617628
Change-Id: I8f07033ea20343cb69055e8b0e17a5f3fe8956f1