deblock: Change the plane ordering
For each 64x64 block, filter the Y, U and V planes instead of
filtering the entire Y plane first followed by the entire U and V
planes.

This lets us reuse some of the filtering parameters for the U and
V planes without having to recompute them.

Also removes a lot of branches in the Y path because of no
subsampling.

PiperOrigin-RevId: 311262375
Change-Id: Ifa894fd423a468e1b8d6b4c14354d323927e7fc9