Update assertion on bottom_border in Tile ctor
CL 297902031 (Eliminate YuvBuffer::ShiftBuffer) changed the way PostFilter
shifted the YuvBuffer, but it did not update the assertions in the Tile
constructor related to the borders of the YuvBuffer. After reviewing the
assertions, I found that the first one does not need to be changed (because it
checks the sum of the left and right borders, which is not changed by the
buffer shift), but the second one should be changed.

Also add assertions to the PostFilter constructor to validate that the
YuvBuffer has enough right and bottom borders to accomodate the horizontal and
vertical shifts.

PiperOrigin-RevId: 314209887
Change-Id: Id1d95b5042682ee6e609cce5a4ae39eae5dc687c