Do not call release_input_buffer if EnqueueFrame fails
Change the API contract so that release_input_buffer is called
only on those frames whose EnqueueFrame calls were successful. If
the Enqueueing fails, libgav1 will not make use of the data so
there is no need for the application to keep it alive.

There are subtler elements to this CL:
 * If ParseAndSchedule fails, the temporal_unit will no longer be
   held in temporal_units_.
 * As a result, ParseAndSchedule will now be simply called and the
   value will be returned instead of invoking SignalFailure on its
   return value.
 * Because of the new structure of ParseAndSchedule,
   EncodedFrame.temporal_unit can no longer be set when it is being
   constructed as we only have a local variable "TemporalUnit" at
   that point. Instead, we set it after the local variable is moved
   into the |temporal_units_| queue.

PiperOrigin-RevId: 314772036
Change-Id: I1d9469bf5a11dc1536d550dad8d0b03bc390bb92