Initialize the reference_frame_index array to -1
The = {} initializer that we use for the ObuFrameHeader struct
initializes the struct members to 0 by default. Since 0 is a valid value
for reference_frame_index, this makes it harder to detect the use of
logically-uninitialized reference_frame_index values for intra frames.
Initialize the reference_frame_index array to -1 instead.
PiperOrigin-RevId: 321050086
Change-Id: I6723cc98d9e22a7457b98f0a41e541b271be2583