Skip MarkInvalidReferenceFrames if show key frame
For a key frame with show_frame equal to 1, we set the entire
decoder_state_.reference_valid array to false. Since the side effect of
MarkInvalidReferenceFrames is that some elements of the
decoder_state_.reference_valid array are set to false, we can skip
MarkInvalidReferenceFrames for a key frame with show_frame equal to 1.

PiperOrigin-RevId: 321060520
Change-Id: Id61ece5b1d41578218b28be4005376be5a265c43