Check kStatusNothingToDequeue only once
Restructure the code that checks the return value of
Decoder::DequeueFrame() so that it checks kStatusNothingToDequeue only
once.

As a byproduct, decoder_fuzzer_frame_parallel.cc no longer depends on
the behavior that Decoder::DequeueFrame() sets *out_ptr to nullptr when
it returns kStatusNothingToDequeue. Note: This behavior is documented,
so it is fine to depend on it.

PiperOrigin-RevId: 321611614
Change-Id: I59f0af293fc3d2a5c152aaf56d30c865676fb167