dsp: update ComputeSuperRes() dsp functions
Extract the filter coefficients loading code of each row. They are stored in
FrameScratchBuffer::superres_coefficients. This buffer is calculated in
SuperResCoefficients_SSE4_1() and SuperResCoefficients_NEON().
They don't have identical results, since NEON does transpose before store.

If the upscale ratio is a constant for all frames, we may create a "global"
buffer for the whole video sequence and only compute it once. However, since it
only occupies ~0.01% CPU time, it's OK to recalculate it for every frame with
upscaling.

For SSE4.1, upscale coefficients are converted to negative values, so that the
maximum coefficient 128 converts to -128 and is within 8-bit signed integer.
Then we can do 8-bit multiplication and accumulation.

Process 16 pixels in each horizontal loop instead of 8.

Decoder speed up for a 1080p sequence with super res on:
~8.2% on an linux machine.
~5.5% on an Android phone.

PiperOrigin-RevId: 326697770
PiperOrigin-RevId: 327089741
Change-Id: I1dd7df5763e425fa92bf4cdf314177a43da192f7