tile: Eliminate the clamp to 3 in ReadCoeffBase*
The clamp to 3 is necessary only if the quantized value was read
using ReadCoeffBaseRange. Add another buffer to store the level
values before ReadCoeffBaseRange is called. The new buffer will
always contain values <= 3 and the std::min calls are no longer
necessary (the context can be computed as a simple addition).

PiperOrigin-RevId: 330981934
Change-Id: I5f441acff53df24cf0be4fb001453fe6c2edcc3f